chore: rename DeleteOrganizationPanel
