+++ packages/frontend/src/components/UserSettings/DeleteOrganizationPanel/DeleteOrganizationPanel.styles.ts
