+++ packages/frontend/src/pages/Settings.tsx
@@ -22,8 +22,8 @@ import PageSpinner from '../components/PageSpinner';
 import AccessTokensPanel from '../components/UserSettings/AccessTokensPanel';
 import AllowedDomainsPanel from '../components/UserSettings/AllowedDomainsPanel';
 import AppearancePanel from '../components/UserSettings/AppearancePanel';
-import { DeleteOrganisationPanel } from '../components/UserSettings/DeleteOrganisationPanel';
-import { Description } from '../components/UserSettings/DeleteOrganisationPanel/DeleteOrganisationPanel.styles';
+import { DeleteOrganizationPanel } from '../components/UserSettings/DeleteOrganizationPanel';
+import { Description } from '../components/UserSettings/DeleteOrganizationPanel/DeleteOrganizationPanel.styles';
 import OrganizationPanel from '../components/UserSettings/OrganizationPanel';
 import PasswordPanel from '../components/UserSettings/PasswordPanel';
 import ProfilePanel from '../components/UserSettings/ProfilePanel';
@@ -222,7 +222,7 @@ const Settings: FC = () => {
                                 </div>
                                 <AllowedDomainsPanel />
                             </CardContainer>
-                            <DeleteOrganisationPanel />
+                            <DeleteOrganizationPanel />
                         </Content>
                     </Route>
                 )}
